+++ govc/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### unreleased
 
+* Add support for local type in the datastore.create command
+
 * Add `-namespace` option to datastore.mkdir and datastore.rm to create/remove namespaces on VSANs
 
 * Add host.service command